requirements.txt
torch
clip-by-openai
Pillow
faiss-cpu
fastapi
uvicorn[standard]
python-multipart